blog can only be deleted by user who created it

+++ controllers/blogs.js
@@ -20,6 +20,7 @@ blogsRouter.post('/', async (request, response, next) => {
     }
 
     const user = await User.findById(decodedToken.id)
+    if (!user) return response.status(401).json({ error: 'cannot find user' })
     
     const blog = new Blog({
       title: body.title,
@@ -40,8 +41,24 @@ blogsRouter.post('/', async (request, response, next) => {
 
 blogsRouter.delete('/:id', async (request, response, next) => {
   try {
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!request.token || !decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+
+    const user = await User.findById(decodedToken.id)
+    if (!user) return response.status(401).json({ error: 'cannot find user' })
+    
+    const removedBlog = await Blog.findById(request.params.id)
+    
+    if ( removedBlog.user.toString() === user._id.toString()) {
+      await Blog.findByIdAndRemove(request.params.id)
+      user.blogs = user.blogs.filter(blog => blog.toString() !== removedBlog._id.toString())
+      await user.save()
+      response.status(204).end()
+    } else {
+      return response.status(401).json({ error: 'You cannot delete blog you did not create' })
+    }
   } catch (exception) {
     next(exception)
   }
